refactor(racingcar-winner): 테스트 구조 변경 및 테스트명 리뷰 반영
