+++ src/test/kotlin/racingcar_winner/CarTest.kt
@@ -11,14 +11,14 @@ class CarTest {
 
     @ParameterizedTest
     @ValueSource(ints = [5, 6, 7, 8, 9])
-    fun `차가 움직일때`(randomNumber: Int) {
+    fun `랜덤 숫자가 4보다 클때 차의 Progress가 증가하는지 확인`(randomNumber: Int) {
         car.tryToMoveForward(randomNumber)
         assertThat(car.progress).isEqualTo(1)
     }
 
     @ParameterizedTest
     @ValueSource(ints = [0, 1, 2, 3, 4])
-    fun `차가 움직이지 않을때`(randomNumber: Int) {
+    fun `랜덤 숫자가 4보다 작을때 차의 Progress가 증가하는지 확인`(randomNumber: Int) {
         car.tryToMoveForward(randomNumber)
         assertThat(car.progress).isEqualTo(0)
     }
